// GUI/src/types/userProfileSettings.ts
export interface UserProfileSettings {
  userId: number;
  forwardedChatPopupNotifications: boolean;
  forwardedChatSoundNotifications: boolean;
  forwardedChatEmailNotifications: boolean;
  newChatPopupNotifications: boolean;
  newChatSoundNotifications: boolean;
  newChatEmailNotifications: boolean;
  useAutocorrect: boolean;
}